+++ src/controls/number.ts
@@ -56,7 +56,20 @@ declare global {
     interface IntrinsicElements {
       /** Form-control with number input
        *  @see {@link WUPNumberControl} */
-      [tagName]: WUP.Number.JSXProps; // add element to tsx/jsx intellisense
+      [tagName]: WUP.Base.ReactHTML<WUPNumberControl> & WUP.Number.JSXProps; // add element to tsx/jsx intellisense (react)
+    }
+  }
+}
+
+// @ts-ignore - because Preact & React can't work together
+declare module "preact/jsx-runtime" {
+  namespace JSX {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    interface HTMLAttributes<RefType> {}
+    interface IntrinsicElements {
+      /** Form-control with number input
+       *  @see {@link WUPNumberControl} */
+      [tagName]: HTMLAttributes<WUPNumberControl> & WUP.Number.JSXProps; // add element to tsx/jsx intellisense (preact)
     }
   }
 }