+++ src/controls/select.ts
@@ -73,11 +73,23 @@ declare global {
     interface IntrinsicElements {
       /** Form-control with dropdown/combobox behavior
        *  @see {@link WUPSelectControl} */
-      [tagName]: WUP.Select.JSXProps; // add element to tsx/jsx intellisense
+      [tagName]: WUP.Base.ReactHTML<WUPSelectControl> & WUP.Select.JSXProps; // add element to tsx/jsx intellisense (react)
     }
   }
 }
 
+// @ts-ignore - because Preact & React can't work together
+declare module "preact/jsx-runtime" {
+  namespace JSX {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    interface HTMLAttributes<RefType> {}
+    interface IntrinsicElements {
+      /** Form-control with dropdown/combobox behavior
+       *  @see {@link WUPSelectControl} */
+      [tagName]: HTMLAttributes<WUPSelectControl> & WUP.Select.JSXProps; // add element to tsx/jsx intellisense (preact)
+    }
+  }
+}
 /** Form-control with dropdown/combobox behavior
  * @example
   const el = document.createElement("wup-select");