+++ src/controls/time.ts
@@ -83,7 +83,20 @@ declare global {
     interface IntrinsicElements {
       /** Form-control with timepicker
        *  @see {@link WUPTimeControl} */
-      [tagName]: WUP.Time.JSXProps; // add element to tsx/jsx intellisense
+      [tagName]: WUP.Base.ReactHTML<WUPTimeControl> & WUP.Time.JSXProps; // add element to tsx/jsx intellisense (react)
+    }
+  }
+}
+
+// @ts-ignore - because Preact & React can't work together
+declare module "preact/jsx-runtime" {
+  namespace JSX {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    interface HTMLAttributes<RefType> {}
+    interface IntrinsicElements {
+      /** Form-control with timepicker
+       *  @see {@link WUPTimeControl} */
+      [tagName]: HTMLAttributes<WUPTimeControl> & WUP.Time.JSXProps; // add element to tsx/jsx intellisense (preact)
     }
   }
 }