+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 <!-- markdownlint-disable MD024 -->
+
 # Changelog
+
+## 1.0.2 (\_\_\_)
+
+### New Features
+
+---
+
+- **Global**
+  - Add Preact tsx/jsx support
+- **Controls**
+  - Added `text-align: start` by default to avoid unexpected inheritance
+  - [RadioControl](src/controls/radio.ts). Added attribute `[checked]` to re-style whole item
+
+---
+
+---
 
 ## 1.0.1 (Nov 17, 2023)